Decompose PDF report context helpers
Split card-building and primary-candidate resolution into focused PDF adapter modules, move report metadata construction next to the report data models, and keep mapping.py on direct helper imports without adding use_cases dependencies.
