+++ docs/ai/repo-map.md
@@ -31,7 +31,7 @@ Scope: single source of truth for detailed file layout, entry points, package st
 - `adapters/http/`: health, clients, settings, recording, history, websocket, updates, car library, and debug route groups; `adapters/http/dependencies.py` owns the grouped router dependency dataclasses and `adapters/http/__init__.py` assembles the router from them.
 - `adapters/websocket/hub.py`: live WebSocket connection fan-out and payload delivery.
 - `adapters/persistence/`: SQLite history DB (`history_db/` keeps `HistoryDB` as the public facade and splits internals across `_run_lifecycle.py`, `_sample_io.py`, `_queries.py`, `_schema.py`, and `_samples.py`) and the static car library loader (`car_library.py`).
-- `adapters/pdf/`: PDF/report rendering pipeline and report mapping entrypoints. `report_context.py` owns context assembly (data-prep, card-assembly, domain aggregate reconstruction) and bridges `use_cases/history/report_interpretation`; `mapping.py` is the thin mapper (context → `ReportTemplateData`) and does not import from `use_cases/`; `pdf_page1.py` and `pdf_page2.py` are thin page composers over focused `_panel_*.py` section modules.
+- `adapters/pdf/`: PDF/report rendering pipeline and report mapping entrypoints. `report_context.py` now focuses on `ReportMappingContext` plus summary/data-prep bridge work, `_candidate_resolver.py` owns primary-candidate resolution, `_card_builder.py` owns system-card assembly/signature humanization, and `report_data.py` owns `Report` plus renderer-facing dataclasses; `mapping.py` remains the thin mapper (context → `ReportTemplateData`) and does not import from `use_cases/`; `pdf_page1.py` and `pdf_page2.py` are thin page composers over focused `_panel_*.py` section modules.
 - `adapters/udp/`, `adapters/gps/`, `adapters/simulator/`, `adapters/hotspot/`: UDP protocol transport, GPS speed ingestion, simulator tooling, and hotspot/AP operational adapters. `adapters/simulator/` now splits pure frame generation into `sim_client.py`, pure road-scene mutation into `sim_scene.py`, asyncio UDP/runtime loops into `sim_runtime.py`, and leaves `sim_sender.py` as the CLI entry/orchestration layer.
 - `infra/runtime/`: lifecycle management, processing loop, runtime health state, and WebSocket broadcast coordination. `registry.py` owns raw `ClientRegistry` state tracking plus `client_snapshots()`, while `client_snapshot.py` owns the client-row presenter reused by `/api/clients` and `ws_broadcast.py`.
 - `infra/processing/`: signal processing pipeline (buffers, FFT, payload shaping, and processor facade).
